minor refactoring and found that box2d caused some building errors

+++ src/kNgine/Camera.hpp
@@ -24,7 +24,7 @@ namespace kNgine
       DIAGONAL//idk what I will use this for
     } fovType = MAX_WH;
     bool showDebugHitBox = false;
-    struct{u64*engineObjectLength;EngineObject**engineObjects;}engineInfo;
+    struct{size_t*engineObjectLength;EngineObject**engineObjects;}engineInfo;
     Camera();
     Camera(f32 fov, u32 windowWidth, u32 windowHeight);
     ~Camera();
